[CS] A simpler base
async-interceptor is reduced to fewer assumptions. It deals only in full
pedestal contexts, and waits to take from the response channel until its
leave function.

We lose the niceness of the IntoInterceptor protocol, but the
`interceptor` provides different niceities: the response channel is
added to the context during the enter function, so further interceptors
may use it if needed.