[WM SD] Allow setting timeouts for each interceptor
We were running into the hard-coded 10 second timeout tripping us up
when testing some upstream timeout behavior. Obviously we'd never
tolerate timeouts this long in production, but we were trying to get a
successful fallback response by setting the HTTP API gateway timeout
higher and weren't able to go above 10 seconds because of this.

This now allows optionally setting the timeout when you create a bifrost
interceptor and makes it clear that the timeout is coming from bifrost
in the response body when it occurs.